+++ crates/onboarding/src/multibuffer_hint.rs
@@ -172,7 +172,7 @@ impl Render for MultibufferHint {
                                 .color(Color::Muted),
                         )
                         .on_click(move |_event, _, cx| {
-                            cx.open_url("https://codeberg.org/ZZZEditor/ZZZ/wiki/multibuffers")
+                            cx.open_url("https://github.com/Xero-Team/ZZZ/wiki/multibuffers")
                         }),
                     ),
             )