+++ crates/open_router/src/open_router.rs
@@ -521,7 +521,7 @@ pub async fn stream_completion(
         .uri(uri)
         .header("Content-Type", "application/json")
         .header("Authorization", format!("Bearer {}", api_key))
-        .header("HTTP-Referer", "https://codeberg.org/ZZZEditor/ZZZ")
+        .header("HTTP-Referer", "https://github.com/Xero-Team/ZZZ")
         .header("X-Title", "ZZZ")
         .extra_headers(extra_headers)
         .body(AsyncBody::from(
@@ -614,7 +614,7 @@ pub async fn list_models(
         .uri(uri)
         .header("Accept", "application/json")
         .header("Authorization", format!("Bearer {}", api_key))
-        .header("HTTP-Referer", "https://codeberg.org/ZZZEditor/ZZZ")
+        .header("HTTP-Referer", "https://github.com/Xero-Team/ZZZ")
         .header("X-Title", "ZZZ")
         .extra_headers(extra_headers)
         .body(AsyncBody::default())