+++ crates/theme_selector/src/icon_theme_selector.rs
@@ -340,7 +340,7 @@ impl PickerDelegate for IconThemeSelectorDelegate {
                             .color(Color::Muted),
                     )
                     .on_click(|_event, _window, cx| {
-                        cx.open_url("https://codeberg.org/ZZZEditor/ZZZ");
+                        cx.open_url("https://github.com/Xero-Team/ZZZ");
                     }),
                 )
                 .child(