+++ crates/theme_selector/src/theme_selector.rs
@@ -542,7 +542,7 @@ impl PickerDelegate for ThemeSelectorDelegate {
                             .color(Color::Muted),
                     )
                     .on_click(cx.listener(|_, _, _, cx| {
-                        cx.open_url("https://codeberg.org/ZZZEditor/ZZZ");
+                        cx.open_url("https://github.com/Xero-Team/ZZZ");
                     })),
                 )
                 .child(